backend-python/requirements.txt
# Web Framework & Server (Fast & Async)
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4

# Environment Variables
python-dotenv==1.0.1

# AI & LLM Orchestration (Switched to Groq for ultra-low latency)
langchain==0.1.13
langchain-groq==0.1.3

# Voice Telemetry & Audio Processing
librosa==0.10.1          # For deep audio/vocal stress analysis
soundfile==0.12.1        # Required by librosa to read audio files
numpy==1.26.4            # For math operations on audio arrays

# PDF Generation
reportlab==4.1.0         # Fast and robust PDF generation from text/images

# Parental Triage / Notifications
twilio==9.0.2            # For sending WhatsApp alerts
emails==0.6              # For sending automated emails

# Async Utilities & File Handling
aiofiles==23.2.1         # For handling file uploads (audio) asynchronously
python-multipart==0.0.9  # Required by FastAPI to accept file uploads from frontend